refactor: router 수정
1. preview-okr 제거
2. main layout 폴더 위치 수정
3. sing in page 공통 layout 제거

related to: #15
