+++ src/Router.tsx
@@ -1,6 +1,5 @@
 import {createBrowserRouter}  from 'react-router-dom';
 import Error from '@components/Error';
-import MainLayout from '@components/MainLayout';
 import AddOkr from './AddOkr';
 import History from './History';
 import Home from './Home';
@@ -8,11 +7,12 @@ import MainDashBoard from './MainDashBoard';
 import PreviewOkr from './PreviewOkr';
 import SignIn from './SignIn';
 import Social from './Social';
+import MainLayout from '@components/layout/MainLayout';
 
 const router = createBrowserRouter([
   {
     path: '/',
-    element: <MainLayout />,
+    element: <MainLayout/>,
     errorElement: <Error />,
     children: [
       {
@@ -32,19 +32,15 @@ const router = createBrowserRouter([
         element: <MainDashBoard />
       },
       {
-        path: '/preview-okr',
-        element: <PreviewOkr />
-      },
-      {
-        path: '/sign-in',
-        element: <SignIn />
-      },
-      {
         path: '/social',
         element: <Social />
       },
     ]
-  }
+  },
+  {
+    path: '/sign-in',
+    element: <SignIn />
+  },
 ])
 
 export default router;
